+++ api-core/engine/src/main/java/cn/tongdun/kunpeng/api/engine/model/rule/action/AssignmentAction.java
@@ -40,6 +40,12 @@ public class AssignmentAction implements Action {
             case "context":
                 right = new Field(rightValue, "object");
                 break;
+            case "indicatrix":
+                right = new PlatformIndex(rightValue, false, null);
+                break;
+            case "index":
+                right = new PolicyIndex(rightValue);
+                break;
             default:
                 throw new ParseException("AssignmentAction parse error!");
         }